refactor(SCRUM-156) : createFirstStar 서비스 수정
