+++ src/main/java/com/team_nebula/nebula/domain/AI/dto/GetThumbnailAndKeywordsResponseDTO.java
@@ -1,4 +1,4 @@
-package com.team_nebula.nebula.domain.star.dto.response;
+package com.team_nebula.nebula.domain.AI.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,17 +6,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
-import java.util.UUID;
 
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class CreateStarResponseDTO {
-
-    private UUID starId;
-    private String title;
-    private String siteUrl;
-    private String thumbnailUrl;
+public class GetThumbnailAndKeywordsResponseDTO {
+    private String image_url;
     private List<String> keywords;
 }
